requirements.txt
flask==3.0.3
python-dotenv==1.0.1
deemix==3.6.6
deezer-py==5.8.0
requests==2.32.3
